backend/requirements.txt
requests
google-generativeai
pysui
Flask
gunicorn